stabilized object detection and segmentation, coco2yolo
