+++ oak_detection/models/custom_models/__init__.py
@@ -1,4 +1,4 @@
 import torch
 import torch.nn as nn
-import oak_object_detection.models.custom_models.YOLOV8.YOLOV8
+import oak_detection.models.custom_models.YOLOV8.YOLOV8
 # import oak_classification.models.custom_models.MOBILENET_V2.MOBILENET_V2
